+++ inlang/source-code/paraglide/paraglide-js-adapter-next/src/app/middleware/resolveLanguage.ts
@@ -1,9 +1,9 @@
 import { NextRequest } from "next/server"
 import { RoutingStragey } from "../routing/interface"
 import { ResolvedI18nConfig } from "../config"
-import { HeaderNames, LANG_COOKIE } from "../constants"
+import { ACCEPT_LANGUAGE_HEADER_NAME, LANG_COOKIE } from "../constants"
 import { isAvailableLanguageTag } from "$paraglide/runtime.js"
-import { preferredLanguages } from "../utils/language"
+import { negotiateLanguagePreferences } from "@inlang/paraglide-js/internal/adapter-utils"
 
 /**
  * Returns the language that should be used for this request
@@ -26,9 +26,7 @@ export function resolveLanguage<T extends string>(
 	const localeCookeValue = request.cookies.get(LANG_COOKIE.name)?.value
 	if (isAvailableLanguageTag(localeCookeValue)) return localeCookeValue as T
 
-	const acceptLanguage = request.headers.get(HeaderNames.AcceptLanguage)
-	if (!acceptLanguage) return config.defaultLanguage
-
-	const preferences = preferredLanguages(acceptLanguage, config.availableLanguageTags)
-	return preferences[0] ?? config.defaultLanguage
+	const acceptLanguage = request.headers.get(ACCEPT_LANGUAGE_HEADER_NAME)
+	const preferences = negotiateLanguagePreferences(acceptLanguage, config.availableLanguageTags)
+	return preferences.at(0) ?? config.defaultLanguage
 }