+++ inlang/source-code/paraglide/paraglide-js-adapter-next/src/app/routing/interface.ts
@@ -1,7 +1,13 @@
 import { NextRequest } from "next/server"
 
 export interface RoutingStragey<T extends string> {
-	getLocalisedPath(canonicalPath: string, locale: T): string
+	getLocalisedPath(
+		canonicalPath: string,
+		locale: T,
+		ctx: {
+			isLocaleSwitch: boolean
+		}
+	): string
 	getCanonicalPath(localisedPath: string, locale: T): string
 	resolveLanguage(request: NextRequest): T | undefined
 	translatePath(localisedPath: string, currentLocale: T, newLocale: T): string