+++ inlang/source-code/paraglide/paraglide-js-adapter-rollup/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @inlang/paraglide-js-adapter-rollup
 
+## 1.0.43
+
+### Patch Changes
+
+- @inlang/paraglide-js-adapter-unplugin@1.4.26
+
+## 1.0.42
+
+### Patch Changes
+
+- @inlang/paraglide-js-adapter-unplugin@1.4.25
+
 ## 1.0.41
 
 ### Patch Changes